Add option to change sigma for UCB
Was previously using a constant (2) for sigma. This works for most use cases but in a particular use case (same model, but one pulls multiple times per one output, the upper confidence was too low for the version that pulled once per output) it's helpful to be able to manually set it.